debut de fix backprop, float pose pb (utiliser double) et la derivee n'est surement pas bonne

+++ train/training_functions.h
@@ -9,5 +9,5 @@ float av_CPDW(struct Network net, struct Neuron* neuron, size_t windex, struct t
 float av_CPDB(struct Network net, struct Neuron* neuron, struct training_set minibatch, float* costs);
 // calculate the average of pd of cost when each w or b changes, and apply wk = wk-nAVPDC(w)  or b = b-nAVPDC(b)
 void back_propagate(struct Network* net, struct training_set minibatch, float rate);
-void train(struct Network* net, struct training_set set, size_t rate, size_t minibatch_size, size_t epochs);
+void train(struct Network* net, struct training_set set, double rate, size_t minibatch_size, size_t epochs);
